Fix Enemy type Close combat
Fixing the pursuit of the player.
Upon requests, a parameter was added where if an enemy entity sees an entity with the Enemy layer at a distance, it refuses to move, falling into a kind of paralysis, but is able to attack, turn around and later move on.
The parameter is called ObstacleDetectionDistance.
For this to work properly, make sure your entities have an Enemy layer.