Update prefabs Enemy
In order for enemies to better switch between ranged and melee choices, each entity needs a projectile targeting.
Attention, in the prefab settings you can see the gizmos rendering which will show and render the targeting area.
For this prefab you need to specify Character in which Combat Mode should be selected. That has been updated to customize gizmos rendering.